Fixed a damage underflow (bugreport:3624)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@14077 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/battle.c
@@ -1840,12 +1840,8 @@ static struct Damage battle_calc_weapon_attack(struct block_list *src,struct blo
 			}
 			if (def1 > 100) def1 = 100;
 			ATK_RATE2(
-				flag.idef ?100:
-				(flag.pdef ?flag.pdef *(def1 + vit_def):
-				100-def1),
-			  	flag.idef2?100:
-				(flag.pdef2?flag.pdef2*(def1 + vit_def):
-				100-def1)
+				flag.idef ?100:(flag.pdef ?(int)(flag.pdef *(def1+vit_def)):(100-def1)),
+			 	flag.idef2?100:(flag.pdef2?(int)(flag.pdef2*(def1+vit_def)):(100-def1))
 			);
 			ATK_ADD2(
 				flag.idef ||flag.pdef ?0:-vit_def,
